+++ src/main/java/org/cloud/sonic/driver/android/service/UiaClient.java
@@ -85,7 +85,7 @@ public interface UiaClient {
 
     void tap(int x, int y) throws SonicRespException;
 
-    void longPress(int x, int y, int ms) throws SonicRespException;
+    void longPress(double x, double y, double ms) throws SonicRespException;
 
     void swipe(int fromX, int fromY, int toX, int toY, Integer duration) throws SonicRespException;
 }
